+++ src/main/java/com/Tretyak_Marina/javacore/chapter10/repository/LabelRepository.java
@@ -2,6 +2,6 @@ package com.Tretyak_Marina.javacore.chapter10.repository;
 
 import com.Tretyak_Marina.javacore.chapter10.model.Label;
 
-public interface LabelRepository extends GenericRepository<Label,Integer> {
+public interface LabelRepository extends GenericRepository<Label,Long> {
 }
