+++ src/main/java/com/Tretyak_Marina/javacore/chapter10/repository/PostRepository.java
@@ -2,6 +2,6 @@ package com.Tretyak_Marina.javacore.chapter10.repository;
 
 import com.Tretyak_Marina.javacore.chapter10.model.Post;
 
-public interface PostRepository extends GenericRepository<Post,Integer> {
+public interface PostRepository extends GenericRepository<Post,Long> {
 }
