+++ src/main/java/com/Tretyak_Marina/javacore/chapter10/repository/WriterRepository.java
@@ -2,6 +2,6 @@ package com.Tretyak_Marina.javacore.chapter10.repository;
 
 import com.Tretyak_Marina.javacore.chapter10.model.Writer;
 
-public interface WriterRepository extends GenericRepository<Writer,Integer> {
+public interface WriterRepository extends GenericRepository<Writer,Long> {
 }
